Add a spec name so we can have multiple databases passed in
This allows us to have more than one cluster in our application and
still test them. FYI for now this won't work with fixtures because we
have to change how the fixtures rake task works. That's a second effort
that I'm also working on, but isn't technically part of the parallel
testing PR.